+++ src/global.js
@@ -10,6 +10,7 @@ var os = require('os');
 var log = require('./helpers/log');
 var program = require('commander');
 var package = require('./helpers/package');
+var config = require('./client/config');
 // var MemoryFS = require("memory-fs");
 
 var __hiipack__ = {
@@ -56,10 +57,11 @@ var __hiipack__ = {
     }
 };
 
-global.__hiipack__ = __hiipack__;
-global.__hii__ = __hiipack__;
 global.log = log;
 global.program = program;
+global.__hiipack__ = __hiipack__;
+global.__hii__ = __hiipack__;
+global.__hii_config__ = config.readFileToObject();
 // global.mfs = new MemoryFS();
 
 module.exports = __hiipack__;