+++ src/scripts/postinstall.js
@@ -13,17 +13,17 @@ var homeDir = os.homedir();
 try{
     var hiipackPath = path.join(homeDir, '.hiipack');
     if(!fs.existsSync(hiipackPath)){
-        fs.mkdir(hiipackPath);
+        fs.mkdirSync(hiipackPath);
     }
 }catch(e){
-    console.error('make dir .hiipack/cache or .hiipack/code failed.', e);
+    console.error('make dir .hiipack/cache or .hiipack/code failed ==>', e.message);
 }
 
 try{
     var hiircPath = path.join(homeDir, '.hiirc');
 
     if(!fs.existsSync(hiircPath)){
-        fs.writeFile(hiircPath, [
+        fs.writeFileSync(hiircPath, [
             'registry=',
             'system_proxy='
         ].join('\n'));
@@ -32,5 +32,5 @@ try{
         console.log(fs.readFileSync(hiircPath, 'utf8'));
     }
 }catch(e){
-    console.error('hiipack config file `.hiirc` create failed: ', e)
+    console.error('hiipack config file `.hiirc` create failed ==>', e.message)
 }