+++ CHANGELOG.md
@@ -8,5 +8,8 @@
 * support open browser firefox/safari/opera
 * support launch firefox with pac proxy
 * support launch opera with pac proxy
+* support dependence version
+
+* pac proxy support system proxy
 
 * fixbug: dll plugin `process.env.NODE_ENV` not right